examples/shv-nxboot-updater/update-script/requirements.txt
pyshv==0.8.0
PyQt6
argparse
asyncio
logging
